Add files via upload
Welcome to the Rock Paper Scissors Game repository! This project is a web-based implementation of the classic game, developed using HTML, CSS, and JavaScript. The goal of this project is to provide a fun and interactive user experience while showcasing fundamental web development skills.
Technologies Used:-
HTML5: Structure and content of the web pages.
CSS3: Styling and layout to create an appealing design.
JavaScript: Game logic, interactivity, and dynamic content updates.